% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyModel.R
\name{applyModel}
\alias{applyModel}
\title{Produce a segmentation based on a given model and extract enhancer / promoter elemenents.}
\usage{
applyModel(rpmCounts, regions, model, genomeSize, outdir = ".",
  nthreads = 1, learnTrans = FALSE)
}
\arguments{
\item{rpmCounts}{Count matrix matching with the \code{regions} parameter.
Each row of the matrix represents a mark and each column a bin resulting
from dividing the genomic regions into non-overlapping bins of equal size.
The rows of the matrix must be named with the name of the marks and these names must be unique.}

\item{regions}{GRanges object containing the genomic regions of interest.}

\item{model}{A list with the parameters that describe the HMM.}

\item{genomeSize}{vector with chromosome lengths.}

\item{outdir}{path to the output directory.}

\item{nthreads}{number of threads used for learning.}

\item{learnTrans}{flag, whether or not to let the model learn the transition probabilities while fixing the emission parameters.}
}
\value{
A list with the following arguments:
}
\description{
Produce a segmentation based on a given model and extract enhancer / promoter elemenents.
}
